+++ src/lib/feature/publicPage/ui/expandableText/ExpandableText.tsx
@@ -17,7 +17,7 @@ export default function ExpandableText({ text, maxLength = 72, className }: Prop
     <div>
       <Typography
         variant={'regular_text_14'}
-        className={twMerge(className, 'inline whitespace-pre-wrap')}
+        className={twMerge(className, 'whitespace-pre-wrap, inline break-all')}
       >
         {displayText}
         {isLong && !expanded && '... '}
